refactor: move token resolution onto the StorageProvider interface
resolve-token.ts is now a thin dispatcher that never names a provider:
resolveOsfToken/resolveGdriveToken move verbatim onto osfProvider/
gdriveProvider as resolveToken(userData, owner), and the dispatcher
delegates through the existing getProviderForExperiment() so the
legacy "no storageProvider means OSF" rule stays in exactly one place.

The public default export keeps its (user_data, exp_data) signature, so
all five callers (api-data, api-base64, scheduled-upload-retry,
create-experiment, get-provider-access-token) are untouched.

Behavior is preserved for an unregistered provider id: getProvider()
throws, but callers distinguish a thrown error (HTTP 500
TOKEN_RESOLUTION_ERROR) from a returned failure (HTTP 400 with a
specific message), so the dispatcher catches the lookup and converts it
back to the original PROVIDER_NOT_CONNECTED result. Only the lookup is
wrapped; resolveToken itself still throws as before.

Dispatching through the registry widens resolve-token's import graph to
both adapters, which surfaced two pre-existing test-environment
fragilities. Fixed in test config only, with no assertion changes:
  - providers-osf, providers-gdrive, put-file-osf-ref now run under
    @jest-environment node. Under the project-default jsdom, jose (via
    firebase-admin/auth) resolves to its ESM-only browser build, which
    Jest's CJS transform cannot parse; the node environment picks its
    CJS build. resolve-token-gdrive.test.js has always carried this
    docblock for the same reason.
  - resolve-token-gdrive mocks "node-fetch" (ESM-only, no CJS build),
    matching what the other three adapter suites already do. It never
    exercises those write paths.

Full suite green: 40 suites / 257 tests, unchanged from baseline.
